Rename app folder to auth
